Refactor subscriptions.json to allow multiple paths
- Allow a Subscription to declare multiple paths that it is subscribed
to.  Whenever any of the files in the paths are changed, the Action the
Subscription references is invoked. This change reduces duplication in
subscriptions.json, because several subscriptions use the same logic for
multiple paths.

- Put arguments for the Action in an "actionArguments" object.  This
removes the need for the built-in Maestro properties (such as "action"
and "delay") to have a prefix, and allows an action argument to have the
same name as a built-in property.

- Detect errors when parsing subscriptions.json by using
JsonPropertyAttribute.Required to ensure all required properties are
present and using JsonSerializerSettings.MissingMemberHandling to ensure
there are no extraneous or misnamed properties.