Revert changes to refactor subscriptions.json
This reverts commits a89a6d79, 25f4f7b4, and c99e4841.

The Maestro web app has not been published to Azure since making these
changes, so the published app is out of sync with the
subscriptions.json in this repo.  With eerhardt out, we need to decide
our plan for publishing Maestro; in the meantime, the changes are being
reverted to get it back in sync.